+++ components/persistor/HISTORY.md
@@ -1,3 +1,5 @@
+## 9.0.0
+* target es2020
 ## 8.0.2
 * bump nconf version due to security vulnerability.
 ## 8.0.1